requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
chromadb==0.4.22
PyPDF2==3.0.1
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
pandas==2.0.3